+++ src/frequenz/gridpool/config/_microgrid.py
@@ -3,7 +3,6 @@
 
 """Data model for microgrid configurations."""
 
-import logging
 import re
 from dataclasses import field
 from datetime import datetime
@@ -11,9 +10,6 @@ from typing import Any, ClassVar, Literal, Self, Type, cast, get_args
 
 from marshmallow import Schema
 from marshmallow_dataclass import dataclass
-
-_logger = logging.getLogger(__name__)
-
 
 ComponentType = Literal["grid", "pv", "battery", "consumption", "chp", "ev"]
 """Valid component types."""
@@ -41,11 +37,6 @@ class ComponentTypeConfig:
     def __post_init__(self) -> None:
         """Set the default formula if none is provided."""
         self.formula = self.formula or {}
-        if "AC_ACTIVE_POWER" in self.formula:
-            _logger.warning(
-                "ComponentTypeConfig: 'AC_ACTIVE_POWER' formula is deprecated, "
-                "please use 'AC_POWER_ACTIVE' instead."
-            )
 
     def cids(self, metric: str = "") -> list[int]:
         """Get component IDs for this component.
@@ -171,9 +162,9 @@ class BatteryConfig:
 
 
 # pylint: disable=too-many-instance-attributes
-@dataclass(frozen=True)
-class Metadata:
-    """Metadata for a microgrid."""
+@dataclass
+class MicrogridConfig:
+    """Configuration of a microgrid."""
 
     microgrid_id: int
     """ID of the microgrid."""
@@ -201,14 +192,6 @@ class Metadata:
 
     end_time: datetime | None = None
     """End time of the microgrid operation."""
-
-
-@dataclass
-class MicrogridConfig:
-    """Configuration of a microgrid."""
-
-    meta: Metadata
-    """Metadata of the microgrid."""
 
     pv: dict[str, PVConfig] | None = None
     """Configuration of the PV system."""
@@ -318,13 +301,9 @@ class MicrogridConfig:
                 raise TypeError("Table reader: Each microgrid entry must be a dict")
 
             mgrid = cls.Schema().load(entry)
-            if mgrid.meta is None or mgrid.meta.microgrid_id is None:
+            if int(mgrid.microgrid_id) != int(mid):
                 raise ValueError(
-                    "Table reader: Each microgrid entry must have a meta.microgrid_id"
-                )
-            if int(mgrid.meta.microgrid_id) != int(mid):
-                raise ValueError(
-                    f"Table reader: Microgrid ID mismatch: key {mid} != {mgrid.meta.microgrid_id}"
+                    f"Table reader: Microgrid ID mismatch: key {mid} != {mgrid.microgrid_id}"
                 )
 
             mgrids[int(mid)] = mgrid